+++ Scraping/Scraping/App.xaml.cs
@@ -5,6 +5,7 @@ using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
+using Scraping.View;
 
 namespace Scraping
 {
