+++ Scraping/Scraping/View/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Scraping.MainWindow"
+﻿<Window x:Class="Scraping.View.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Scraping"
+        xmlns:local="clr-namespace:Scraping.View"
         mc:Ignorable="d"
         Closed="Window_Closed"
         Title="百合漫画　発売日" Height="500" Width="900">
