+++ Scraping/Scraping/View/RegisterDialog.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Scraping.RegisterDialog"
+﻿<Window x:Class="Scraping.View.RegisterDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Scraping"
+        xmlns:local="clr-namespace:Scraping.View"
         mc:Ignorable="d"
         Title="RegisterDialog" Height="250" Width="300" WindowStyle="ToolWindow">
     <Grid>
