+++ Scraping/Scraping/View/RegisterDialog.xaml.cs
@@ -12,8 +12,10 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Scraping.ViewModel;
+using Scraping;
 
-namespace Scraping
+namespace Scraping.View
 {
     /// <summary>
     /// RegisterDialog.xaml の相互作用ロジック
